requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
boto3==1.21.46
botocore==1.24.46
dj-database-url==0.5.0
dj-email-url==1.0.4
Django==4.0.4
django-cache-url==3.2.3
django-ckeditor==6.4.0
django-cors-headers==3.11.0
django-environ==0.8.1
django-extensions==3.1.5
django-js-asset==2.0.0
django-storages==1.12.3
django-taggit==2.1.0
djangorestframework==3.13.1
gunicorn==20.1.0
jmespath==1.0.0
marshmallow==3.15.0
packaging==21.3
Pillow==9.1.0
psycopg2-binary==2.9.3
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.0.0
